Code cleanup

+++ bi_gol_mpp_android/app/src/main/java/software/egger/gol/BoardViewModel.kt
@@ -21,7 +21,7 @@ class BoardViewModel : ViewModel() {
 
     }
 
-    var cellSize: Float? = null
-    var offsetX: Float? = null
-    var offsetY: Float? = null
+    var cellSize: Double? = null
+    var offsetX: Double? = null
+    var offsetY: Double? = null
 }
